Remove dead call code
